requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
openai>=1.0.0
aiohttp>=3.8.0
tenacity>=8.0.0
tqdm>=4.64.0 